refactor: split Controller file
