+++ src/main/kotlin/racingcar/Application.kt
@@ -1,32 +1,20 @@
 package racingcar
 
+import camp.nextstep.edu.missionutils.Randoms
+
+
 fun main() {
-    // TODO: Implement the program
+
 }
 
 
-object InputController {
-    fun readNamesOfCars(): List<String>? {
-        return readlnOrNull()?.split(",")
-    }
-
-    fun readRounds(): Int? {
-        return readlnOrNull()?.toInt()
-    }
-}
 
 
-object OutputController {
-    private const val ENTER_THE_NAMES_OF_CARS_OUTPUT = "Enter the names of the cars (comma-separated):"
-    private const val PLAY_ROUND_OUTPUT = "How many rounds will be played?"
-    private const val RACE_RESULTS = "Race Results"
-    private const val WINNERS = "Winners"
+object RandomNumController {
+    private val startRandomNum = 0
+    private val endRandomNum = 9
 
-    fun printNamesOfCars() {
-        println(ENTER_THE_NAMES_OF_CARS_OUTPUT)
-    }
-
-    fun printPlayerRound() {
-        println(PLAY_ROUND_OUTPUT)
+    fun getRandomNum(): Int {
+        return Randoms.pickNumberInRange(startRandomNum, endRandomNum)
     }
 }
